=== Release 1.70.0 ===
- Bug fixes
- CPU performance improvement when particles become inactive
- Proper calculation of coordination number in Sample Analyzer
- More correct calculation of center of geometries, considering surfaces of triangles
- Fix mass calculation in SampleAnalyzerTab
- Better update strategy for 3D scene.
    - Add interfaces for dialogs to request update particular objects' types on the scene instead of the whole OpenGL view.
    - SampleAnalyzerTab updates only volumes -> much faster on large scenes.
- Extend text export/import. Add possibility to export/import analysis volumes, package generators, bonds generators.
- PackageGenerator with PBC.
    - Consider PBC when calculating number of particles in PackageGenerator - chosen porosity will be properly reached when PCB is enabled.
    - Fix tolerance when checking consistency of analysis volumes.
- Possibility to define contact diameter when importing mixture from text
- Proper import/export of geometries' motion.
- New fracture criteria for bonds
- Add RTX 3090 to CUDA config
- Extend solid bond model with an additional breakage criterion
- Possibility to define time-dependent motion of geometries in cmusen script.
- Fix not working dynamic generator if mixture generation is selected.
- Possibility to define force-dependent motion of geometries in cmusen script.
- Remove "Planes Coordinates" from export as text, use conventional "Coordinates" instead.
- fix: Set tangential overlap in BondsGenerator
- build: Upgrade development environment
    - Visual Studio 2017 -> 2019.
    - Windows Target platform 8 -> 10.
    - C++11 -> C++17.
    - CUDA 10.0 -> 11.2.
    - Qt 5.11.2 -> 5.15.2.
    - Qt installer 3.2.3 -> 4.0.1.
    - Inno Setup 6.0.5 -> 6.1.2.
    - Protobuf 3.9.1 -> 3.14.0.
    - CMake 3.14 -> 3.18.
    - gcc 7 -> 10.
- Proper check internal overlaps in Dynamic Generator
- Improve notifications when opening files in the GUI.
- Application of tangential damping only in the case when no slipping occur

+++ QTDialogs/QtOpenGLView/OpenGLView.h
@@ -125,7 +125,7 @@ protected:
 
 	void SetColorGL(const CColor& _color);
 
-	void GetObjectColor(unsigned int nObjectID, CColor* _ResultColor); // return for specified object ID its colors according to the coloring scheme
+	void GetObjectColor(size_t nObjectID, CColor* _ResultColor); // return for specified object ID its colors according to the coloring scheme
 	CVector3 WinCoord2GL(double _x, double _y, double _z) const;
 
 private:
